requirements.txt
c7n>=0.9.16
c7n-gcp>=0.4.42
c7n-azure>=0.7.0
google-cloud-bigquery>=2.30.0
pandas>=1.3.5
pyyaml>=6.0
azure-mgmt-costmanagement>=1.0.0